Implemented printReport and its test
Implemented the tests after doubling back on my research on how to test
println output. Then implemented the actual printReport function (at
least for the refactorings side of it) via the private function. Made
few mistakes that test caught such as using || when I needed && and when
I used <= instead of < for comparing the counter to nSuggestions.