Made System Test for Report Class
Made the system test for Report class in order to check ouputs and make
sure they are working as expected with what is currently implemented in
regards to the SuggestedRefactorings. System test might be updated with
stuff Isaac is doing given they are also working on Report, but
different part.