Fleshes out PropertiesFile tests more and adds content to files
Files now contain information to pull from and tests are added. Some
tests are still in for consideration as noted in the comments, this is
mainly because I'm unsure if they should throw an error and stop the
program, throw an error and continue working using defaults, or do
nothing and just defaults and so forth.