Updates Report Class to Handle Source Files
In progress of updating Report class to include the source code part of
the report. This involves updating updating methods and units test. Also
involves updating references as the constructor has been changed.

IMPORTANT: Not everything compiles right now in this version. Just committing because need a momentary break to do something else and want to save the work.