Integration Tests Updated and Removed Old Files
I include data from report system test now nto the integration test file
in its own method call and updated it to handle a couple things in
regards to file paths. Would be cool to get a huge list of file paths to
use at some point.

With that said I also removed the ReportSystemTest.java given it doesn't
have use anymore in effort to tidy up the project. I also removed
FilePaths and FilePathsTest because originally it was split off into own
class to handle the source file report part of the class.  However, it
was intended originally for Reports to still call that class and pass
information to it. Ultimately though as I designed the Report class with
that in mind, there was only one method that the ReportClass didn't have
a hand in already which was the sortSourceFiles, so I made the decision
to pull all that stuff into one Object which is the Report class. There
wasn't much in the original FilePaths.java class anyway.