[TOOL] Adds PR task list template
This is in an effort to help us all remember
important steps in a PR. Like we've talked about,
we want to step up our testing game this time around.
Having others review our code is a means to offer
constructive feedback on how things can be improved.

The checklist also helps us each remember what should
be considered when merging.