Updates Supplied Code for Accurate Reflection
The code in the suggested refactoring generation is updated to always
create 100 random suggestions now. Before hand it was using
numSuggestions as the cap, which in demostration purposes would always
cap off the amount printed and not show a feature properly. With this
adjustment, say the user only wants 5 suggestions printed. We can now
see properly on the last line an example of say: "Printed 5 of 81
suggestions." Before hand it would cap out as: "Printed 5 of 5
suggestions."