Adds More System Tests for automation
Probably last bit I can do in the moment. There is probably better way
to test certain features that can be ripped straight from reading the
output string, but my brain isn't up to it at the moment. Was able to
handle a tricky solution in regards dupe files test, but not the others.